% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{annotations-set}
\alias{annotations-set}
\alias{annotations<-}
\alias{annotations<-,epiRomicsS4-method}
\title{Replace the annotations slot of an epiRomicsS4 object}
\usage{
annotations(x, ...) <- value

\S4method{annotations}{epiRomicsS4}(x, ...) <- value
}
\arguments{
\item{x}{An \code{\linkS4class{epiRomicsS4}} object.}

\item{...}{Currently unused; reserved for method extension.}

\item{value}{A \code{\link[GenomicRanges]{GRanges}} object.}
}
\value{
The updated \code{\linkS4class{epiRomicsS4}} object.
}
\description{
Assigns a new \code{GRanges} object to the \code{annotations} slot.
Validity is checked via \code{\link[methods]{validObject}}.
}
\examples{
db <- make_example_database()
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(1, 100))
gr$type <- "hg38_custom_h3k4me1"
annotations(db) <- gr
length(annotations(db))
}
