% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{annotations}
\alias{annotations}
\alias{annotations,epiRomicsS4-method}
\title{Access the genomic annotations slot of an epiRomicsS4 object}
\usage{
annotations(x, ...)

\S4method{annotations}{epiRomicsS4}(x, ...)
}
\arguments{
\item{x}{An \code{\linkS4class{epiRomicsS4}} object.}

\item{...}{Currently unused; reserved for method extension.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object. Empty
  (length 0) if no annotations have been assigned.
}
\description{
Returns the \code{GRanges} object stored in the \code{annotations}
slot. Public accessor replacement for \code{obj@annotations} or
\code{methods::slot(obj, "annotations")}.
}
\examples{
db <- make_example_database()
annotations(db)
length(annotations(db))
}
\seealso{
\code{\link{annotations<-}}, \code{\linkS4class{epiRomicsS4}}
}
