% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-deprecated.R
\name{epiRomics-deprecated}
\alias{epiRomics-deprecated}
\alias{epiRomics_build_dB}
\alias{epiRomics_cache_data}
\alias{epiRomics_cache_path}
\alias{epiRomics_has_cache}
\alias{epiRomics_chromatin_states}
\alias{epiRomics_chromatin_states_categories}
\alias{epiRomics_enhanceosome}
\alias{epiRomics_enhancer_predictor_to_ref}
\alias{epiRomics_enhancers_co_marks}
\alias{epiRomics_enhancers_filter}
\alias{epiRomics_filter_accessible}
\alias{epiRomics_annotate_putative}
\alias{epiRomics_putative_enhancers}
\alias{epiRomics_quick_view}
\alias{epiRomics_regions_of_interest}
\alias{epiRomics_tf_cobinding}
\alias{epiRomics_tf_overlap}
\alias{epiRomics_track_layer}
\alias{epiRomics_track_layer_fast}
\alias{epiRomics_track_layer_gene}
\title{Renamed epiRomics functions}
\usage{
epiRomics_build_dB(...)

epiRomics_cache_data(...)

epiRomics_cache_path(...)

epiRomics_has_cache(...)

epiRomics_chromatin_states(...)

epiRomics_chromatin_states_categories(...)

epiRomics_enhanceosome(...)

epiRomics_enhancer_predictor_to_ref(...)

epiRomics_enhancers_co_marks(...)

epiRomics_enhancers_filter(...)

epiRomics_filter_accessible(...)

epiRomics_annotate_putative(...)

epiRomics_putative_enhancers(...)

epiRomics_quick_view(...)

epiRomics_regions_of_interest(...)

epiRomics_tf_cobinding(...)

epiRomics_tf_overlap(...)

epiRomics_track_layer(...)

epiRomics_track_layer_fast(...)

epiRomics_track_layer_gene(...)
}
\arguments{
\item{...}{Arguments forwarded to the replacement function.}
}
\description{
These functions have been renamed in epiRomics 0.99.1 to follow
Bioconductor naming conventions. They remain as thin wrappers that
emit a one-time rename notice and forward to the new name. They
will be removed in the next release.
}
\keyword{internal}
