% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_enhancers.R
\name{filter_enhancers}
\alias{filter_enhancers}
\title{Filters putative enhancers called by
epiRomics_enhancers by crossing against curated
FANTOM data}
\usage{
filter_enhancers(putative_enhancers, database, type = NULL)
}
\arguments{
\item{putative_enhancers}{epiRomics class
database containing putative enhancer calls}

\item{database}{epiRomics class database
containing all data initially loaded}

\item{type}{epiRomics reference containing
database to validate putative enhancers against.
Default NULL derives from genome:
\code{paste0(genome, "_custom_fantom")}.}
}
\value{
Variable of class epiRomics with filtered candidate enhancer regions
}
\description{
Filters putative enhancers called by
epiRomics_enhancers by crossing against curated
FANTOM data
}
\examples{
db <- make_example_database()
eso <- make_example_enhanceosome(db)
## filter_enhancers requires fantom annotations in database
tryCatch(
  filter_enhancers(eso, db),
  error = function(e) message(e$message)
)
}
