% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanceosomes.R
\name{find_enhanceosomes}
\alias{find_enhanceosomes}
\title{Identifies putative enhanceosome regions by
cross-referencing candidate enhancer regions
against co-TF enrichment}
\usage{
find_enhanceosomes(putative_enhancers, database)
}
\arguments{
\item{putative_enhancers}{epiRomics class
database containing putative enhancer calls}

\item{database}{epiRomics class database
containing all data initially loaded}
}
\value{
Variable of class epiRomics with enhanceosome annotations
}
\description{
Identifies putative enhanceosome regions by
cross-referencing candidate enhancer regions
against co-TF enrichment
}
\examples{
db <- make_example_database()
eso <- make_example_enhanceosome(db)
length(methods::slot(eso, "annotations"))
}
