% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancers.R
\name{find_enhancers_by_comarks}
\alias{find_enhancers_by_comarks}
\title{Identifies putative enhancer regions utilizing
select histone mark co-occurrence}
\usage{
find_enhancers_by_comarks(
  database,
  histone_mark_1 = "h3k4me1",
  histone_mark_2 = "h3k27ac"
)
}
\arguments{
\item{database}{epiRomics class database
containing all data initially loaded}

\item{histone_mark_1}{name of first
histone mark, must match a \code{name} entry in
\code{\link{meta}(database)}, default set to h3k4me1}

\item{histone_mark_2}{name of second
histone mark, must match a \code{name} entry in
\code{\link{meta}(database)}, default set to h3k27ac}
}
\value{
Variable of class epiRomics further
  exploring candidate enhancer regions identified
  after histone integration
}
\description{
Identifies putative enhancer regions utilizing
select histone mark co-occurrence
}
\examples{
db <- make_example_database()
enhancers <- find_enhancers_by_comarks(db)
length(annotations(enhancers))
}
