% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{genome-set}
\alias{genome-set}
\alias{genome<-}
\alias{genome<-,epiRomicsS4-method}
\title{Replace the genome assembly name}
\usage{
\S4method{genome}{epiRomicsS4}(x) <- value
}
\arguments{
\item{x}{An \code{\linkS4class{epiRomicsS4}} object.}

\item{value}{A non-empty character scalar.}
}
\value{
The updated \code{\linkS4class{epiRomicsS4}} object.
}
\description{
Assigns a new genome assembly name (e.g. \code{"hg38"}, \code{"mm10"})
to the \code{genome} slot. This method extends the
\code{genome<-()} replacement generic from the \pkg{GenomeInfoDb}
package.
}
\examples{
db <- make_example_database()
genome(db) <- "hg38"
genome(db)
}
