% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{meta-set}
\alias{meta-set}
\alias{meta<-}
\alias{meta<-,epiRomicsS4-method}
\title{Replace the metadata slot of an epiRomicsS4 object}
\usage{
meta(x, ...) <- value

\S4method{meta}{epiRomicsS4}(x, ...) <- value
}
\arguments{
\item{x}{An \code{\linkS4class{epiRomicsS4}} object.}

\item{...}{Currently unused; reserved for method extension.}

\item{value}{A \code{data.frame}.}
}
\value{
The updated \code{\linkS4class{epiRomicsS4}} object.
}
\description{
Assigns a new \code{data.frame} to the \code{meta} slot.
}
\examples{
db <- make_example_database()
new_meta <- meta(db)
new_meta$extra <- "annotated"
meta(db) <- new_meta
colnames(meta(db))
}
