% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{organism-set}
\alias{organism-set}
\alias{organism<-}
\alias{organism<-,epiRomicsS4-method}
\title{Replace the organism annotation package name}
\usage{
\S4method{organism}{epiRomicsS4}(object) <- value
}
\arguments{
\item{object}{An \code{\linkS4class{epiRomicsS4}} object.}

\item{value}{A character scalar, typically an
\code{org.*.eg.db} package name.}
}
\value{
The updated \code{\linkS4class{epiRomicsS4}} object.
}
\description{
Assigns a new \code{org.*.eg.db} package name to the \code{organism}
slot. This method extends the \code{organism<-()} replacement
generic from the \pkg{BiocGenerics} package.
}
\examples{
db <- make_example_database()
organism(db) <- "org.Hs.eg.db"
organism(db)
}
