% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{organism}
\alias{organism}
\alias{organism,epiRomicsS4-method}
\title{Access the organism annotation package name}
\usage{
\S4method{organism}{epiRomicsS4}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{epiRomicsS4}} object.}
}
\value{
A character scalar. Empty character if unset.
}
\description{
Returns the character string stored in the \code{organism} slot
(e.g. \code{"org.Hs.eg.db"}, \code{"org.Mm.eg.db"}). This method
extends the \code{organism()} generic from the
\pkg{BiocGenerics} package.
}
\examples{
db <- make_example_database()
organism(db)
}
