% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{txdb-set}
\alias{txdb-set}
\alias{txdb<-}
\alias{txdb<-,epiRomicsS4-method}
\title{Replace the TxDb identifier slot of an epiRomicsS4 object}
\usage{
txdb(x, ...) <- value

\S4method{txdb}{epiRomicsS4}(x, ...) <- value
}
\arguments{
\item{x}{An \code{\linkS4class{epiRomicsS4}} object.}

\item{...}{Currently unused; reserved for method extension.}

\item{value}{A character scalar in
\code{"TxDb.Package::TxDb.object"} form.}
}
\value{
The updated \code{\linkS4class{epiRomicsS4}} object.
}
\description{
Assigns a new \code{"TxDb.Package::TxDb.object"} character string to
the \code{txdb} slot. Validity is checked via
\code{\link[methods]{validObject}}.
}
\examples{
db <- make_example_database()
txdb(db) <- "TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene"
txdb(db)
}
