% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{txdb}
\alias{txdb}
\alias{txdb,epiRomicsS4-method}
\title{Access the TxDb package::object identifier slot}
\usage{
txdb(x, ...)

\S4method{txdb}{epiRomicsS4}(x, ...)
}
\arguments{
\item{x}{An \code{\linkS4class{epiRomicsS4}} object.}

\item{...}{Currently unused; reserved for method extension.}
}
\value{
A character scalar. Empty character if unset.
}
\description{
Returns the character string stored in the \code{txdb} slot. The value
has the form \code{"TxDb.Package::TxDb.object"}, e.g.
\code{"TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene"}.
Public accessor replacement for \code{obj@txdb} or
\code{methods::slot(obj, "txdb")}.
}
\examples{
db <- make_example_database()
txdb(db)
}
\seealso{
\code{\link{txdb<-}}, \code{\linkS4class{epiRomicsS4}}
}
