% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics-package.R
\docType{package}
\name{epiRomics-package}
\alias{epiRomics}
\alias{epiRomics-package}
\title{epiRomics: Epigenomic Analysis Package Built for R (epiRomics)}
\description{
Integrates various levels of epigenomic information, including ChIP-seq, histone modification, ATAC-seq, and RNA-seq data. Regulatory network analysis uses combinatory approaches to infer regions of significance, such as enhancers. Downstream analysis identifies co-occurrence of epigenomic data at regions of interest. Visualization functions display multi-track genomic views with signal overlays. Please contact <ammawla@ucdavis.edu> for suggestions, feedback, or bug reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Huising-Lab/epiRomics}
  \item Report bugs at \url{https://github.com/Huising-Lab/epiRomics/issues}
}

}
\author{
\strong{Maintainer}: Alex M. Mawla \email{ammawla@ucdavis.edu} (\href{https://orcid.org/0000-0003-0907-464X}{ORCID})

Authors:
\itemize{
  \item Mark O. Huising \email{mhuising@ucdavis.edu} (\href{https://orcid.org/0000-0002-6594-2205}{ORCID})
}

Other contributors:
\itemize{
  \item Huising Lab @ University of California, Davis [copyright holder]
  \item National Institutes of Health NIDDK (NIDDK110276 to M.O.H.) [funder]
  \item Juvenile Diabetes Research Foundation (CDA-2-2013-54 and 2-SRA-2021-1054-M-N to M.O.H.) [funder]
  \item Stephen F. and Bettina A. Sims Immunology Fellowship (Fellowship to A.M.M.) [funder]
  \item Amazon Web Services (Machine Learning Research Award to A.M.M.) [funder]
  \item UC Davis Flow Cytometry Shared Resource Laboratory (NCI P30 CA093373 and NIH S10 OD018223) [funder]
  \item University of California, Davis (Start-up funds to T.v.d.M.) [funder]
}

}
