% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_build_dB_v2.R
\name{epiRomics_build_dB_2}
\alias{epiRomics_build_dB_2}
\title{Build epiRomics database}
\usage{
epiRomics_build_dB_2(
  epiRomics_db_file,
  txdb_organism,
  epiRomics_genome,
  epiRomics_organism
)
}
\arguments{
\item{epiRomics_db_file}{character string of path to properly formatted csv file containing epigenetic data. [See vignette for more details]}

\item{txdb_organism}{a character string containing the TxDB associated with your data.}

\item{epiRomics_genome}{a character string containing the genome associated with your data. e.g. "mm10" or "hg38".}

\item{epiRomics_organism}{a character string containing the org.db associated with your data.}
}
\value{
Variable of class epiRomics for further downstream analysis
}
\description{
Build epiRomics database
}
