% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_chromatin_tracks.R
\name{epiRomics_chromatin_tracks}
\alias{epiRomics_chromatin_tracks}
\title{Visualizes chromatin availability}
\usage{
epiRomics_chromatin_tracks(
  epiRomics_gene_name,
  epiRomics_dB,
  epiRomics_track_connection
)
}
\arguments{
\item{epiRomics_gene_name}{character of name of gene to visualize}

\item{epiRomics_dB}{epiRomics class database containing all data initially loaded}

\item{epiRomics_track_connection}{data frame containing bigwig track locations and their names}
}
\value{
GViz plot
}
\description{
Visualizes chromatin availability
}
