% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_enhanceosome.R
\name{epiRomics_enhanceosome}
\alias{epiRomics_enhanceosome}
\title{Identifies putative enhanceosome regions by cross-referencing candidate enhancer regions against co-TF enrichment}
\usage{
epiRomics_enhanceosome(epiRomics_putative_enhancers, epiRomics_dB)
}
\arguments{
\item{epiRomics_putative_enhancers}{epiRomics class database containing putative enhancer calls}

\item{epiRomics_dB}{epiRomics class database containing all data initially loaded}
}
\value{
Variable of class epiRomics further exploring candidate enhanceosome regions using co-ChIP hits
}
\description{
Identifies putative enhanceosome regions by cross-referencing candidate enhancer regions against co-TF enrichment
}
