% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_enhancer_predictor_test.R
\name{epiRomics_enhancer_predictor_test}
\alias{epiRomics_enhancer_predictor_test}
\title{Interrogates various histone marks against a curated database to determine which are most informative}
\usage{
epiRomics_enhancer_predictor_test(
  epiRomics_dB,
  epiRomics_histone = "h3k4me1",
  epiRomics_curated_database = "fantom"
)
}
\arguments{
\item{epiRomics_dB}{epiRomics class database containing all data initially loaded}

\item{epiRomics_histone}{name or vector of histone mark(s), must match name in epiROmics_dB@meta, default set to h3k4me1}

\item{epiRomics_curated_database}{database to test histone marks against, must match name in epiROmics_dB@meta default set to fantom}
}
\value{
Variable of class dataframe further exploring top histone marks that may determine enhancer regions
}
\description{
Interrogates various histone marks against a curated database to determine which are most informative
}
