% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_enhancers.R
\name{epiRomics_enhancers}
\alias{epiRomics_enhancers}
\title{Identifies putative enhancer regions utilizing select histone marks}
\usage{
epiRomics_enhancers(
  epiRomics_dB,
  epiRomics_histone_mark_1 = "h3k4me1",
  epiRomics_histone_mark_2 = "h3k27ac"
)
}
\arguments{
\item{epiRomics_dB}{epiRomics class database containing all data initially loaded}

\item{epiRomics_histone_mark_1}{name of first histone mark, must match name in epiROmics_dB@meta, default set to h3k4me1}

\item{epiRomics_histone_mark_2}{name of second histone mark, must match name in epiROmics_dB@meta default set to h3k27ac}
}
\value{
Variable of class epiRomics further exploring candidate enhancer regions identified after histone integration
}
\description{
Identifies putative enhancer regions utilizing select histone marks
}
