% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_enhancers_filter.R
\name{epiRomics_enhancers_filter}
\alias{epiRomics_enhancers_filter}
\title{Filters putative enhancers called by epiRomics_enhancers by crossing against curated FANTOM data}
\usage{
epiRomics_enhancers_filter(
  epiRomics_putative_enhancers,
  epiRomics_dB,
  epiRomics_type = "mm10_custom_fantom"
)
}
\arguments{
\item{epiRomics_putative_enhancers}{epiRomics class database containing putative enhancer calls}

\item{epiRomics_dB}{epiRomics class database containing all data initially loaded}

\item{epiRomics_type}{epiRomics reference containing database to validate putative enhancers against}
}
\value{
Variable of class epiRomics with filtered candidate enhancer regions
}
\description{
Filters putative enhancers called by epiRomics_enhancers by crossing against curated FANTOM data
}
