% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_predictors.R
\name{epiRomics_predictors}
\alias{epiRomics_predictors}
\title{Predicts TF behavior in association with enhanceosome presence in accordance with functional annotation}
\usage{
epiRomics_predictors(epiRomics_putative_enhanceosome)
}
\arguments{
\item{epiRomics_putative_enhanceosome}{epiRomics class database containing putative enhanceosome calls}
}
\value{
Returned decision tree available for plotting
}
\description{
Predicts TF behavior in association with enhanceosome presence in accordance with functional annotation
}
