% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_region_tracks.R
\name{epiRomics_region_tracks}
\alias{epiRomics_region_tracks}
\title{Visualizes chromatin availability at custom regions}
\usage{
epiRomics_region_tracks(
  epiRomics_region,
  epiRomics_dB,
  epiRomics_track_connection
)
}
\arguments{
\item{epiRomics_region}{GRanges of region to visualize}

\item{epiRomics_dB}{epiRomics class database containing all data initially loaded}

\item{epiRomics_track_connection}{data frame containing bigwig track locations and their names}
}
\value{
GViz plot
}
\description{
Visualizes chromatin availability at custom regions
}
