% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiRomics_regions_of_interest.R
\name{epiRomics_regions_of_interest}
\alias{epiRomics_regions_of_interest}
\title{Evaluates whether regions of interest derived from external experiments, such as ATAC-Seq, correspond with enhanceosome regions}
\usage{
epiRomics_regions_of_interest(
  epiRomics_putative_enhanceosome,
  epiRomics_test_regions
)
}
\arguments{
\item{epiRomics_putative_enhanceosome}{epiRomics class database containing putative enhanceosome calls}

\item{epiRomics_test_regions}{GRanges containing regions of interest}
}
\value{
Variable of class epiRomics with enhanceosome regions overlapping with regions of interest
}
\description{
Evaluates whether regions of interest derived from external experiments, such as ATAC-Seq, correspond with enhanceosome regions
}
