% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-epiRomics.R
\docType{class}
\name{epiRomicsS4-class}
\alias{epiRomicsS4-class}
\alias{epiRomicsS4}
\title{An S4 class to manage epiRomics databases and downstream results}
\description{
An S4 class to manage epiRomics databases and downstream results
}
\section{Slots}{

\describe{
\item{\code{annotations}}{GRanges}

\item{\code{meta}}{data.frame}

\item{\code{txdb}}{txdb string name}

\item{\code{organism}}{org.db string name}

\item{\code{genome}}{genome name, e.g. 'mm10' or 'hg19'}
}}

